requirements.txt
PySide6>=6.7
SQLAlchemy>=2.0
watchdog>=4.0
Pillow>=10.0
python-dotenv>=1.0
